// frontend/src/types/technologyWeek.ts
export interface TechnologyWeek {
    id: number
    kvk_id: number
    kvk_name: string
    reporting_year: string
    start_date: string
    end_date: string
    type_of_activity: string
    no_of_activities: number
    related_technology: string
    no_of_participants: number
    created_at?: string
    updated_at?: string
}